requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
streamlit
requests
pandas
plotly
fpdf
pyzbar
Pillow
opencv-python-headless